Added a page where a new volunteer objects can be viewed
Added a new volunteer class that has a list with all the charity events it is participating in. Modified FileContentProvider to be able to read and write said volunteers. Added a page that has a list of all saved volunteers with the option of viewing the charity events they are working on. Also added a new saved file with three volunteers for future testing.