Added use of aggregate (#84)
* Added use of aggregate

Minor fixes and added a use of aggregate to find which volunteer has the most equipment.
Fixed pagination. Made the list of potential events show only events that you can add the volunteer into, eliminating empty choices when managing a volunteer.
Now when choosing potential events for volunteers the user is presented with short secondary lines describing how much the event needs this new volunteer.
Fixed line alignment and naming.
A new service is here to help with event separation.